Fix Travis config (#1)
* Actually install Tox on Travis

* Properly configure Tesseract

* Quit using tox to manage tests

* Compile a recent Tesseract manually on Travis

Because "yeah, distros never update anything, here's some prebuilt Linux
binaries for you!" would be too hard.

* Follow curl redirects, and only run on 3.5 for now

* Potentially autotool the crap out of leptonica

This might be the first time I've tried to use a software package that
didn't give me a ./configure script.

* Maybe autotools heard me whining and broke Travis

* Maybe I should be using my Linux VM for this

* Apparently the VM was overkill

* Don't run more commands than needed

Apparently "just run everything" was a common enough use case that the
`autoreconf` command already does that. Who knew?

* Fine, you can run as root

* Cache things

* Use a more recent Leptonica version

* Do a bunch of smarter things

* Manually link against leptonica???

I don't know if this will do this, or if this is why nothing works, but
I no longer know anything.

* Hopefully, skip compilation if it's unnecessary

* Wind up in the right place after compiling tesseract

* Why can't anything be simple or intuitive?

* Potentially improve lep/tess handling

This means I have to say things in the actual commit log
if I want to rebuild things on Travis, but I'm OK with that.

* Apparently you need this

* Be significantly smarter about builds and installs

* Fix a permissions thing

* Look in the right place for compiled Tesseract

* Pretend to compile tesseract if not actually compiling

Apparently tesseract doesn't stay compiled so `sudo make install` was
recompiling it. That's not a good thing.